feature.Refactor components structure
